% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcov.R
\name{haff_eig}
\alias{haff_eig}
\title{Stein/Haff's ordered eigenvalue estimates}
\usage{
haff_eig(l, n)
}
\arguments{
\item{l}{Sample eigenvalues}

\item{n}{Number of observations}
}
\value{
Estimated eigenvalues
}
\description{
Stein/Haff's ordered eigenvalue estimates
}
\examples{
p <- 5
n <- 10
S <- rWishart(1, n, diag(p))[,,1]
l <- eigen(S)$val
haff_eig(l, n)
}

