% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcov.R
\name{iso_cov}
\alias{iso_cov}
\title{Stein's isotonized covariance estimator}
\usage{
iso_cov(S, n)
}
\arguments{
\item{S}{Sample covariance matrix}

\item{n}{Number of observations}
}
\value{
Estimated covariance matrix
}
\description{
Stein's isotonized covariance estimator
}
\examples{
p <- 5
n <- 10
S <- rWishart(1, n, diag(p))[,,1]
iso_cov(S, n)
}

