% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcov.R
\name{iso_eig}
\alias{iso_eig}
\title{Stein's isotonized eigenvalue estimates}
\usage{
iso_eig(l, n)
}
\arguments{
\item{l}{Sample eigenvalues}

\item{n}{Number of observations}
}
\value{
Estimated eigenvalues
}
\description{
Stein's isotonized eigenvalue estimates
}
\examples{
p <- 5
n <- 10
S <- rWishart(1, n, diag(p))[,,1]
l <- eigen(S)$val
iso_eig(l, n)
}

